backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
livekit-api==1.1.0
